Move 'pretty' to its own module
