+++ servant-docs/src/Servant/Docs.hs
@@ -53,4 +53,5 @@ module Servant.Docs
   , single
   ) where
 
-import           Servant.Docs.Internal
+import Servant.Docs.Internal
+import Servant.Docs.Internal.Pretty
